specify scala bundle version
scala 2.9.3 seems not to be packaged with OSGi metadata, so BND cannot
infer the right version by himself. Using the `Require-Bundle`
directive allows us to say which version of the scala library is
required. It is pretty safe to use this directive here because you
will never execute a scala program without a scala library in your OSGi
environment