add a way to retrieve content encoding
By default HTTP 1.1 says that the content is encoded in ISO-8859-1. But
this can be changed by setting the charset in the parameters of the
`Content-Type` header

This data is available in the request header (available e.g. when
handling multipart post request) and in the general request data
(available through the `talk` object in an `HLet`).

closes #6