[new #9] add some logs for internal stuffs
As it was pointed out to me, tiscaf could be more verbose when the
request is rejected for some reason. This introduces a small and simple
logging API dedicated to tiscaf internals only that tries to log
messages that can provide user with some helpful insights on what's
going on.

For example, if the upload post/put/patch data size was exceeded...

closes #9